+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/activity/AnimeActivityScreen.kt
@@ -2,18 +2,9 @@ package com.thekeeperofpie.artistalleydatabase.anime.activity
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.pullrefresh.PullRefreshIndicator
-import androidx.compose.material.pullrefresh.pullRefresh
-import androidx.compose.material.pullrefresh.rememberPullRefreshState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Tab
 import androidx.compose.material3.TabRow
@@ -23,26 +14,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.input.nestedscroll.nestedScroll
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.paging.LoadState
-import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
-import androidx.paging.compose.itemContentType
-import androidx.paging.compose.itemKey
-import com.anilist.UserSocialActivityQuery.Data.Page.ListActivityActivity
-import com.anilist.UserSocialActivityQuery.Data.Page.MessageActivityActivity
-import com.anilist.UserSocialActivityQuery.Data.Page.OtherActivity
-import com.anilist.UserSocialActivityQuery.Data.Page.TextActivityActivity
-import com.thekeeperofpie.artistalleydatabase.anilist.oauth.AniListViewer
 import com.thekeeperofpie.artistalleydatabase.anime.AnimeNavDestinations
 import com.thekeeperofpie.artistalleydatabase.anime.LocalNavigationCallback
 import com.thekeeperofpie.artistalleydatabase.anime.R
-import com.thekeeperofpie.artistalleydatabase.anime.media.AnimeMediaListScreen
 import com.thekeeperofpie.artistalleydatabase.anime.media.MediaUtils.primaryTitle
 import com.thekeeperofpie.artistalleydatabase.anime.media.edit.MediaEditBottomSheetScaffold
 import com.thekeeperofpie.artistalleydatabase.anime.media.edit.MediaEditViewModel
@@ -52,10 +31,7 @@ import com.thekeeperofpie.artistalleydatabase.compose.EnterAlwaysTopAppBarHeight
 import com.thekeeperofpie.artistalleydatabase.compose.UpIconOption
 import kotlinx.coroutines.launch
 
-@OptIn(
-    ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class,
-    ExperimentalMaterialApi::class
-)
+@OptIn(ExperimentalMaterial3Api::class, ExperimentalFoundationApi::class)
 object AnimeActivityScreen {
 
     private val SCREEN_KEY = AnimeNavDestinations.ACTIVITY.id
@@ -158,119 +134,13 @@ object AnimeActivityScreen {
                         else -> throw IllegalArgumentException("Invalid page")
                     }
                     ActivityList(
+                        screenKey = SCREEN_KEY,
                         editViewModel = editViewModel,
                         viewer = viewer,
                         activities = activities,
                         onActivityStatusUpdate = viewModel.activityToggleHelper::toggle,
                         showMedia = selectedMedia == null,
                     )
-                }
-            }
-        }
-    }
-
-    @Composable
-    private fun ActivityList(
-        editViewModel: MediaEditViewModel,
-        viewer: AniListViewer?,
-        activities: LazyPagingItems<AnimeActivityViewModel.ActivityEntry>,
-        onActivityStatusUpdate: (ActivityToggleUpdate) -> Unit,
-        showMedia: Boolean,
-    ) {
-        when (val refreshState = activities.loadState.refresh) {
-            is LoadState.Error -> AnimeMediaListScreen.Error(
-                exception = refreshState.error,
-            )
-            else -> {
-                if (activities.itemCount == 0
-                    && activities.loadState.refresh is LoadState.NotLoading
-                ) {
-                    AnimeMediaListScreen.NoResults()
-                } else {
-                    // TODO: Move this up a level
-                    val refreshing = activities.loadState.refresh is LoadState.Loading
-                    val pullRefreshState = rememberPullRefreshState(
-                        refreshing = refreshing,
-                        onRefresh = { activities.refresh() },
-                    )
-                    Box(modifier = Modifier.fillMaxWidth()) {
-                        LazyColumn(
-                            contentPadding = PaddingValues(
-                                start = 16.dp,
-                                end = 16.dp,
-                                top = 16.dp,
-                                bottom = 72.dp,
-                            ),
-                            verticalArrangement = Arrangement.spacedBy(16.dp),
-                            modifier = Modifier.pullRefresh(state = pullRefreshState)
-                        ) {
-                            items(
-                                count = activities.itemCount,
-                                key = activities.itemKey { it.activityId.scopedId },
-                                contentType = activities.itemContentType { it.activityId.type }
-                            ) {
-                                val entry = activities[it]
-                                when (val activity = entry?.activity) {
-                                    is TextActivityActivity -> TextActivitySmallCard(
-                                        screenKey = SCREEN_KEY,
-                                        viewer = viewer,
-                                        activity = activity,
-                                        entry = entry,
-                                        onActivityStatusUpdate = onActivityStatusUpdate,
-                                        clickable = true,
-                                        modifier = Modifier.fillMaxWidth()
-                                    )
-                                    is ListActivityActivity -> ListActivitySmallCard(
-                                        screenKey = SCREEN_KEY,
-                                        viewer = viewer,
-                                        activity = activity,
-                                        mediaEntry = entry.media,
-                                        showMedia = showMedia,
-                                        entry = entry,
-                                        onActivityStatusUpdate = onActivityStatusUpdate,
-                                        onClickListEdit = { editViewModel.initialize(it.media) },
-                                        clickable = true,
-                                        modifier = Modifier.fillMaxWidth()
-                                    )
-                                    is MessageActivityActivity -> MessageActivitySmallCard(
-                                        screenKey = SCREEN_KEY,
-                                        viewer = viewer,
-                                        activity = activity,
-                                        entry = entry,
-                                        onActivityStatusUpdate = onActivityStatusUpdate,
-                                        clickable = true,
-                                        modifier = Modifier.fillMaxWidth()
-                                    )
-                                    is OtherActivity,
-                                    null,
-                                    -> TextActivitySmallCard(
-                                        screenKey = SCREEN_KEY,
-                                        activity = null,
-                                        viewer = viewer,
-                                        entry = null,
-                                        onActivityStatusUpdate = onActivityStatusUpdate,
-                                        modifier = Modifier.fillMaxWidth()
-                                    )
-                                }
-                            }
-
-                            when (activities.loadState.append) {
-                                is LoadState.Loading -> item("load_more_append") {
-                                    AnimeMediaListScreen.LoadingMore()
-                                }
-                                is LoadState.Error -> item("load_more_error") {
-                                    AnimeMediaListScreen.AppendError { activities.retry() }
-                                }
-                                is LoadState.NotLoading -> Unit
-                            }
-                        }
-
-                        PullRefreshIndicator(
-                            refreshing = refreshing,
-                            state = pullRefreshState,
-                            modifier = Modifier.align(Alignment.TopCenter)
-                        )
-                    }
                 }
             }
         }