+++ modules/anime/src/main/kotlin/com/thekeeperofpie/artistalleydatabase/anime/user/UserActivityScreen.kt
@@ -1,0 +1,33 @@
+package com.thekeeperofpie.artistalleydatabase.anime.user
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.runtime.Composable
+import androidx.paging.compose.collectAsLazyPagingItems
+import com.thekeeperofpie.artistalleydatabase.anilist.oauth.AniListViewer
+import com.thekeeperofpie.artistalleydatabase.anime.activity.ActivityList
+import com.thekeeperofpie.artistalleydatabase.anime.media.edit.MediaEditViewModel
+import com.thekeeperofpie.artistalleydatabase.anime.media.filter.SortFilterBottomScaffold
+
+@OptIn(ExperimentalMaterial3Api::class)
+object UserActivityScreen {
+
+    @Composable
+    operator fun invoke(
+        viewModel: AniListUserViewModel,
+        editViewModel: MediaEditViewModel,
+        viewer: AniListViewer?,
+    ) {
+        val sortFilterController = viewModel.activitySortFilterController
+        SortFilterBottomScaffold(sortFilterController = sortFilterController) {
+            ActivityList(
+                screenKey = viewModel.screenKey,
+                editViewModel = editViewModel,
+                viewer = viewer,
+                activities = viewModel.activities.collectAsLazyPagingItems(),
+                onActivityStatusUpdate = viewModel.activityToggleHelper::toggle,
+                showMedia = true,
+                allowUserClick = false,
+            )
+        }
+    }
+}